Upgrade react-dropzone to hook-based API
